Return None for non-UTF-8 tags
This prevents a panic when opening a file which contains non-UTF-8 tags,
but still prevents users from accessing them.

In the future it might be useful to provide an API giving the raw bytes
if the user wants to try guessing it, or maybe even add an (optional)
encoding crate dependency for formats which declare their encoding.

Fixes #6.