Add ability to get and set offset into Sound and Music
This feature got a bit mixed up in general bug fixes for
the buffered audio aspect of how music plays, but it is how it is.

So far this appears to have fixed all the edge cases (looping very
small files, getting unlucky when looping right at the end of a
file when buffers are nearly exhausted).